% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/gene_retriever.r
\name{gene_retriever}
\alias{gene_retriever}
\title{GeneRetriever}
\usage{
gene_retriever(gene_names, nrow = 3, dir = "./", csv.out = "output.csv",
  gr_name = "gr_output.pdf", w = 8, h = 11, pdf = TRUE, heatmap = 0,
  hw = 15, hh = 15)
}
\description{
This function formats cuffdiff format gene expression data into a table and plot the data as box plots or heatmaps to provide a easy to read visual representation of the data.
param gene_names This first argument is a vector of gene names (as string),
param nrow The number of rows in boxplot array
}
\examples{
gene_retriever()
}

